Marriage_Prediction/requirements.txt
fastapi
uvicorn
openai
python-dotenv
jinja2